backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
sqlalchemy
python-dotenv
passlib[bcrypt]
python-jose
google-api-python-client
google-auth
google-auth-oauthlib
requests
